RequestAPI, Pokemon
-App Transport Security policy

If you’re developing a new app, you should use
HTTPS exclusively. If you have an existing app,
you should use HTTPS as much as you can right now,
and create a plan for migrating the rest of your
app as soon as possible. In addition, your communication
through higher-level APIs needs to be encrypted using
TLS version 1.2 with forward secrecy.
If you try to make a connection that doesn't follow
this requirement, an error is thrown.
If your app needs to make a request to an insecure domain,
you have to specify this domain in your app's Info.plist file.

-Pokemon model
-requestAPI