autoConstraints: go back to the suggested
The question then is which view a given constraint should belong to. The answer is:
the view that is closest up the view hierarchy from both views involved in the constraint.
If possible, it should be one of those views. If the constraint dictates a view’s
absolute width, it belongs to that view; if it sets the top of a view in relation to the top
of its superview, it belongs to that superview; if it aligns the tops of two sibling views,
it belongs to their common superview